+++ src/game/condition.cpp
@@ -73,11 +73,11 @@ bool BuildingCondition::is_true() const {
 
 bool ComputerCondition::is_true() const {
     if (line < 0) {
-        return op_eq(op, g.mini.currentScreen, static_cast<int>(screen));
+        return op_eq(op, g.mini.currentScreen, screen);
     } else {
         return op_eq(
                 op, std::make_pair(g.mini.currentScreen, g.mini.selectLine),
-                std::make_pair(static_cast<int>(screen), line));
+                std::make_pair(screen, line));
     }
 }
